feat: moved content from statements-functions.ts to base.ts for better consistency
